refactor(mobile): split email store
- extract email capture state module
- add store state helper tests
